925389b insights: add file exists check (#5026)
https://github.com/BuilderIO/qwik/commit/925389b418a58eecc43620251d8268b8084b5325

+++ optimizer.cjs
@@ -3042,9 +3042,11 @@ globalThis.qwikOptimizer = function(module) {
           const fs = await sys.dynamicImport("node:fs");
           try {
             const INSIGHTS_Q_MANIFEST_FILENAME = "./dist/q-insights.json";
-            const entryStrategy = JSON.parse(await fs.promises.readFile(INSIGHTS_Q_MANIFEST_FILENAME, "utf-8"));
-            entryStrategy && (qwikViteOpts.entryStrategy = entryStrategy);
-            await fs.promises.unlink(INSIGHTS_Q_MANIFEST_FILENAME);
+            if (fs.existsSync(INSIGHTS_Q_MANIFEST_FILENAME)) {
+              const entryStrategy = JSON.parse(await fs.promises.readFile(INSIGHTS_Q_MANIFEST_FILENAME, "utf-8"));
+              entryStrategy && (qwikViteOpts.entryStrategy = entryStrategy);
+              await fs.promises.unlink(INSIGHTS_Q_MANIFEST_FILENAME);
+            }
           } catch (e) {}
         }
         "serve" === viteCommand ? qwikViteOpts.entryStrategy = {
